Move subscription API to separate file `clientcubscription.go` for easy maintenance
